backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
sse-starlette>=1.8.0
pydantic>=2.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
